Update build pipeline
