+++ package.json
@@ -15,10 +15,9 @@
     "format:check": "prettier --check {src,test}/*",
     "test": "jest",
     "prebuild": "rimraf dist && mkdir dist",
-    "build": "npm run build:declaration && npm run build:esm && npm run build:cjs",
-    "build:declaration": "tsc --declaration --emitDeclarationOnly --outDir dist",
-    "build:esm": "tsc --module es2015 --target es5 --outDir dist/esm",
-    "build:cjs": "tsc --module commonjs --target es5 --outDir dist/cjs",
+    "build": "npm run build:types && npm run build:lib",
+    "build:types": "tsc --declaration --emitDeclarationOnly --outDir dist",
+    "build:lib": "rollup -c",
     "postversion": "git push && git push --tags"
   },
   "repository": {
@@ -40,12 +39,16 @@
     "@babel/core": "^7.13.10",
     "@babel/preset-env": "^7.13.10",
     "@babel/preset-typescript": "^7.13.0",
+    "@rollup/plugin-typescript": "^8.2.0",
     "@types/jest": "^26.0.20",
     "babel-jest": "^26.6.3",
     "husky": "^5.1.3",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
+    "rollup": "^2.41.2",
+    "rollup-plugin-terser": "^7.0.2",
+    "tslib": "^2.1.0",
     "typescript": "^4.2.3"
   }
 }
