+++ rollup.config.js
@@ -1,0 +1,22 @@
+import typescript from "@rollup/plugin-typescript";
+import { terser } from "rollup-plugin-terser";
+
+export default [
+  {
+    input: "src/index.ts",
+    output: {
+      file: "dist/index.cjs.js",
+      format: "cjs",
+      exports: "default",
+    },
+    plugins: [typescript(), terser()],
+  },
+  {
+    input: "src/index.ts",
+    output: {
+      file: "dist/index.esm.js",
+      format: "es",
+    },
+    plugins: [typescript(), terser()],
+  },
+];
