+++ src/chars.ts
@@ -1,0 +1,11 @@
+// List of the characters we want to escape and their HTML escaped version
+const chars: Record<string, string> = {
+  "&": "&amp;",
+  ">": "&gt;",
+  "<": "&lt;",
+  '"': "&quot;",
+  "'": "&#39;",
+  "`": "&#96;",
+};
+
+export default chars;
